chore: refactor elasticsearch client - expect generateElasticsearchQuery as param
